requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
async-generator=1.10=pypi_0
attrs=22.1.0=pypi_0
biopython=1.78=py39h1a28f6b_0
blas=1.0=openblas
bottleneck=1.3.5=py39heec5a64_0
brotli-python=1.0.9=py39hc377ac9_7
ca-certificates=2023.12.12=hca03da5_0
certifi=2024.2.2=py39hca03da5_0
cffi=1.16.0=py39h80987f9_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
conda=22.9.0=py39hca03da5_0
conda-package-handling=2.2.0=py39hca03da5_0
conda-package-streaming=0.9.0=py39hca03da5_0
cryptography=41.0.7=py39hd4332d6_0
cyrus-sasl=2.1.28=h9131b1a_1
exceptiongroup=1.0.0rc9=pypi_0
gettext=0.21.0=h13f89a0_1
giflib=5.2.1=h80987f9_3
glib=2.69.1=h514c7bf_2
gst-plugins-base=1.14.1=h313beb8_1
gstreamer=1.14.1=h80987f9_1
h11=0.14.0=pypi_0
icu=68.1=hc377ac9_0
idna=3.4=py39hca03da5_0
jpeg=9e=h80987f9_1
krb5=1.20.1=hf3e1bf2_1
lerc=3.0=hc377ac9_0
libclang=14.0.6=default_h1b80db6_1
libclang13=14.0.6=default_h24352ff_1
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20230828=h80987f9_0
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libllvm14=14.0.6=h7ec7a93_3
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libpq=12.17=h02f6b3c_0
libtiff=4.5.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.4=h372ba2a_0
libxslt=1.1.37=habca612_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
mysql=5.7.24=ha71a6ea_2
ncurses=6.4=h313beb8_0
numexpr=2.8.7=py39hecc3335_0
numpy=1.26.3=py39h3b2db8e_0
numpy-base=1.26.3=py39ha9811e2_0
openssl=3.0.13=h1a28f6b_0
outcome=1.2.0=pypi_0
packaging=23.1=py39hca03da5_0
pandas=2.1.4=py39h46d7db6_0
pcre=8.45=hc377ac9_0
pip=23.3.1=py39hca03da5_0
ply=3.11=py39hca03da5_0
pycosat=0.6.6=py39h80987f9_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py39hca03da5_0
pyqt=5.15.7=py39hc377ac9_0
pyqt5-sip=12.11.0=py39hc377ac9_0
pysocks=1.7.1=py39hca03da5_0
python=3.9.18=hb885b13_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python.app=3=py39h1a28f6b_0
pytz=2023.3.post1=py39hca03da5_0
qt-main=5.15.2=h9b4df51_9
qt-webengine=5.15.9=h2903aaf_7
qtwebkit=5.212=h19f419d_5
readline=8.2=h1a28f6b_0
requests=2.31.0=py39hca03da5_0
ruamel_yaml=0.15.100=py39h1a28f6b_0
selenium=4.5.0=pypi_0
setuptools=68.2.2=py39hca03da5_0
sip=6.6.2=py39hc377ac9_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.41.2=h80987f9_0
tk=8.6.12=hb8d0fd4_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py39hca03da5_0
trio=0.22.0=pypi_0
trio-websocket=0.9.2=pypi_0
tzdata=2023d=h04d1e81_0
urllib3=1.26.18=py39hca03da5_0
wheel=0.41.2=py39hca03da5_0
wsproto=1.2.0=pypi_0
xz=5.4.5=h80987f9_0
yaml=0.2.5=h1a28f6b_0
zlib=1.2.13=h5a0b063_0
zstandard=0.19.0=py39h80987f9_0
zstd=1.5.5=hd90d995_0
